#22 #21 #18 #6 added better settings loading, added textures and rendering

+++ src/vector2.cpp
@@ -1,6 +1,6 @@
 #include "vector2.h"
 
-Vector2::Vector2(int16_t x, int16_t y) {
+Vector2::Vector2(int32_t x, int32_t y) {
     X = x;
     Y = y;
 }
@@ -9,24 +9,24 @@ Vector2::Vector2(){
 
 }
 
-int16_t Vector2::getX() {
+int32_t Vector2::getX() {
     return X;
 }
 
-int16_t Vector2::getY() {
+int32_t Vector2::getY() {
     return Y;
 }
 
-void Vector2::setX(int16_t x) {
+void Vector2::setX(int32_t x) {
     X = x;
 }
 
-int32_t Vector2::getCombined() {
-    int32_t combined = X << 8;
+int64_t Vector2::getCombined() {
+    int64_t combined = X << 8;
     return combined + Y;
 }
 
-void Vector2::setY(int16_t y) {
+void Vector2::setY(int32_t y) {
     Y = y;
 }
